backend/services/incident_detection/requirements.txt
fastapi
uvicorn
pymongo
kafka-python
redis
python-dotenv
